add snapshot functionality
This does two things:
- It allows creating snapshots from within the plugin (tested with
  the RAM checkbox disabled). The snapshot is created on the node the
  command was issued (i.e., clicked in the GUI). Make sure this node has a
  regular/diskfull assignment.
- Deleting snapshots.

Currently "rollback" is not possible, as there is a difference between
what proxmox expects and how snapshot restore is handled by drbdmanage.
For now, if everything goes wrong and you need the snapshot, you have to
use the command line drbdmanage client and issue a "drbdmanage
restore-snapshot" command.