switch from DBUS API to CLI
Due to the improved cluster wide locking, nodes, especially satellites,
sometimes need to retry their actions. That is intended. Currently, the
plugin did not handle these retry actions. So "actions" like creating a
resource and then "immediately" listing volumes failed as the leader was
still in its transaction and hold the corresponding lock.

Instead of hacking that into the soon deprecated DBUS API, I switched
the plugin to use the command line client. For sure the CLI is the one
that is tested best, and this strategy works very well for the
OpenNebula plugin, which also calls the DM CLI.