fix waiting until a resource is ready
Using "dd" was a crude hack we inherited from the original plugin
developed at Proxmox, and we kept it for too long. Using "dd" on its own
does not make sense, worse, it does not capture all cases. A DRBD device
can exist and can be readable, still it is not ready to use. Fix that by
using "drbdsetup wait-connect-resource".