explicit storpoolname related improvements for status and list_images
Rename the $json (which actually should have been named $perl,
because it was the decoded json already) to the more appropriate
$r_list, $rd_list, $sp_list.

Make use of $cache in status.

In list_images, query both "resource-definition list" (mostly for the
intended volume size) and "resource list" (for the stor_pool_name
it actually ended up in), to be able to accurately list the images
for a specific storagepool name.