fix activate/deactivate volume to not accidentally remove images
We use the activate/deactive volume hooks to create and delete
intentionally diskless clients for resources on nodes that
do not have a local backend for the resource in question.

The logic to detect whether a particular DRBD resource does exist,
and if it is "intentionally" or "accidentally" Diskless,
or even Unconfigured, was broken.

activate_volume() would try to re-create resources as diskless clients
if they happen to be currently unconfigured, which would then fail.

deactivate_volume() would (hopefully only try to) *delete* resources
and their backends, if they happen to be currently detached.

That should now both be fixed.