fix regression that could remove ALL volumes on ONE vm destruction
regression introduced in
 b0f2d16 explicit storpoolname related improvements for status and list_images
I accidentally dropped the filter on vmid or vollist.

Depending on whether or not Storage/Plugin.pm or Storage.pm itself
applied that same filter after getting the list from the actual plugin,
this could lead to removal of *all* LINSTOR volumes,
because proxmox would call vdisk_free on the returned list,
which was supposed to be filtered by VMID (but was not).