wait-connect-resource: handle 2 node clusters
This function was introduces because there has been a race between
creating the resource and returning. There have been checks (dd'ing from
the device) which are not sufficient for various reasons. Handling that
by wait-connect-resource is better.

Unfortunately this broke 2 node clusters where one half is gone. In such
a scenario drbdsetup will wait forever.

Fix this by:
- reducing the timeout (it was meant for races, it can be shorter).
- catch the timeout exception and check if someone is UpToDate. If that
is the case (for all volumes), we are fine, else die.