switch to REST-API
This switches from calling the commandline client to using LINSTOR's
rest API, which has several advantages:

- The commandline client's machinereadable output is not always
stable/might be removed in the future.
- It was never meant to be consumed by users and information is needed
to be looked up at several places.

The most important change is that the whole code is now split into
modules. The plugin is just a pretty thin wrapper around two modules:

- Linstor.pm: handles all state information and all LINSTOR events like
creation of resources.
- PluginHelper.pm: contains code of functions that do filtering/more
complicated logic.

The nice thing is that we can unit test these modules, which makes
development ways easier.