fix custom diskless pools
LINSTOR reports diskless pools as having infinite storage. This is done
by setting a MAX value. This confuses Proxmox, and in fact we should
report them as using/providing 0 space. The consequence is that such a
node's pool is shown as "inactive", which I think is fine.

As a side note: in 99.9% of all cases uses should not set a specific
diskless pool, but they do. For whatever reason. The default one should
be good enough. The only reason to have one is to set some specific
properties.

fixes #30