create: create storage locally if possible
What this does is to try to create the DRBD resource diskfully if it is
possible on the node where the VM will be executed.

A common setup might consist of 3 nodes, where all have a given storage
pool, with a replica count of 2. By just auto-placing the resource, it
might happened that the storage is placed on nodes A and C, but the VM is
actually executed on B. This is fine, the plugin creates a diskless DRBD
resource on B, and all data is read and written via the network. Still,
it is obviously preferable to have local storage.

What happens is that the plugin tries to create a diskful resource on
the node that will run the VM (with the same storage pool as defined in
the RG).

If this fails: Ignore it, run an auto-place (as we always did). It might
find another placement.
if this succeeds: Cool, still run auto-place, we might need to create
other replicas on other nodes to fulfil the replica count.

As this is a new behavior (i.e., storage is no longer equally
distributed, but the local node is prefered), this behavior has to be
enabled as follows:

drbd: drbdboth
   content images,rootdir
   controller 192.168.122.94
   resourcegroup onboth
   preferlocal yes

This feature is very much coupled to LINSTOR RGs. As this is the
preferred method to configure DRBD/LINSTOR storage for some time, this
commit also removes the old legacy options (setting redundancy and
storage pools). These (and all other LINSTOR/DRBD) options have to be
set on RG level.