snapshot: add rollback
This adds rollback to the *last* snapshot. Rolling back to only the last
snapshot taken is currently a limitation of LINSTOR. AFAIK this is
because of how parent/child relationship in ZFS works. LINSTOR will gain
some means to be more flexible eventually (e.g., delete in-between
snapshots).

Rolling back to 'B', in a snapshot series of 'A', 'B', 'C' currently
requires the user to delete 'C' first.