delete tmp res on snapshot_delete
When a backup is created via a snapshot we create a temporary resource
that is then used to take the actual backup. Unfortunately this
temporary resource was not that temporary...

As plugin we can not really differentiate of a snapshot_delete is an
"ordinary" snapshot delete, or one after a backup. So we can check for
the name and if it matches "vzdump", then we delete the temporary
resource with that name as well.

So what about ordinary snapshots named "vzdump"? Fortunately enough
"vzdump" is a reserved name, so this should be fine.