config: allow multiple controllers
This allows users to set multiple, ','-separated controllers, like
"192.168.122.2,192.168.122.3". The first one that successfully answers a
GET to '/health' is then used.

This is not useful for most users right now, as one still needs to keep
the LINSTOR controller DB highly-available and needs to start/stop
services and mounts. But this will be possible in the future, when we
introduce such an easy to use "cluster manager" specific for that task.
Right now it is a preparation step.