status: add extra cache
For whatever reason Proxmox queries the status of all storages every few
seconds. There is a cache they pass in, but that only makes sense for
multiple storages of one storage type (i.e., 'drbd'). One can then use
that passed in cache for a "query loop" of all storages of that
particular type. That helps, but still the plugin's status function gets
called every few seconds with a cold cache.

On large clusters (~10 nodes, a few hundred resources), executing the
expensive "storage-pool list" query in LINSTOR every few seconds on all
nodes saturates the LINSTOR REST-API/controller and the cluster becomes
slow/unusable.

As the plugin gets instantiated over and over again, it is not possible
to keep a "plugin global" cache, so moving the cache to the file system
looks like a good fit.

The new option 'statuscache' can be set to a positive number of seconds
and will then keep a status cache (currently only the storage pools) for
that number of seconds. If one wants to make use of this, the option has
to be set on *all* drbd storages in /etc/pve/storage.cfg.