rm controllervm setting
This is from a time where we thought managing a HA LINSTOR controller in
a separate VM that is itself managed by PVE could be a good idea. Since
then lessons got learned and the way to go is drbd-reactor::promoter to
setup a HA LINSTOR controller.

This removes a lot of special cases from the code.