keep status cache for 1min by default
This was implemented for a customer back then and we had a user on
github that had that problem as well. I think it exists long enough to
make that the new default.

Between implementing that cache and now LINSTOR now also has an internal
cache but the user (i.e., the plugin) has to decide when to query the
cache and when not by adding a URL parameter. This is nice, but does not
really help, the code would basically be the same (i.e., stat()-ing some
file) to decide when to send the parameter and when not. We have
something that works, and Perl's Storable is easy to use, so keep it.
Probably that is even faster than to do the http call all the times, but
sometimes with the cache=true parameter...