container clone: only create if necessary; rm on snapshot rm
When one clones a container from a snapshot, PVE does:
- activate_volume() a new LINSTOR resource based on the snapshot
  (snap_$snapname...).
- create the new image vm-$id... (alloc_image()).
- internally copies the content from that "temporary" snapshot resource
  to the new image.

This had several problems:
- if a second clone was created we failed as the "temporary" snapshot
  DRBD resource already existed.
- we never removed that "temporary" resource.

The first part is easy, only create that LINSTOR/DRBD resource when
needed, otherwise skip it, that resource is supposed to be read-only.

Deleting is more involved as to my best knowledge we don't get any
"deactivate" call to delete that resource after the copy finished. The
best we can do is to keep that resource and delete it on
volume_snapshot_delete(). We saw that problem already when implementing
vzdump support. I think we can assume that we always want to delete that
snapshot resource when the actual snapshot is removed.

This means that between the clone(s) and the time the base snapshot gets
removed, we keep that temporary/unnecessary resource. But IMO that is
the best we can do and certainly better than not deleting it at all.