status: show cluster wide view
So far we more or less (size reporting was always a mess) tried to show
a node-local view of storage. For example we only showed pools that
actually exist on a node. This was okay, but had two problems:

- showing only the local storage in a distributed cluster does not make
  too much sense. The actual data lands on nodes LINSTOR decides.
- there are people that have the storage nodes pretty much separated
  from the PVE nodes. For these we did not show any storage information
  on such nodes (as they did not have the RG's SP deployed), which is
  rather confusing.

Now we shifted the calculation of free/used storage to LINSTOR and that
is what we show on all nodes.

tl;tr: we shifted from a node local view to a cluster view.

For PVE nodes we have the following assumption:
- linstor-satellite (and all that comes with it: drbd-utils,
  drbd-dkms,...)
- linstor-proxmox plugin installed