fix performance regression in status calls
This was introduced in e3b49eeb52ded60add7f3ef0e5e05a9b6a88ded7

Long story short, PVE calls status() freaking often. Depending on if you
select the storage in the GUI, this can happen every second, otherwise
like every 10s. To accommodate for that Proxmox's status() API has a
"cache" object that is passed in for every storage type. Let's say you
have 4 'drbd' storages (as in storage.cfg), then for the first call we
get an empty cache which we can fill, and for the next 3 we can use
that.

That helps, but
- we can not make use of that for more than one iteration (remember, PVE
  calls status an awful lot), that is why we introduced a file based
  persistent cache to keep some information longer than 10 seconds.
- things get even worse if you have a distributed system where then
  every PVE node hammers the LINSTOR controller with (the same)
  questions all the time.

Do the math, the LINSTOR controller gets too many "what is the status of
my storage" calls very very quickly even in setups with a handful of
nodes.

Recently we improved how we display storage information, but that also
rendered the built in PVE cache useless as we needed to query every RG
individually. That was known, I thought we could get away with it. We
did not.

How was it fixed:
- we can now query the storage information for all RGs once. That helps
  to again use the builtin cache more effectively as we again need 1
  call for every iteration.
- The new LINSTOR API for querying all RGs at once also got a cache
  parameter which also helps a lot because it avoids the pressure on the
  controller that resulted from the distributed nature of LINSTOR/DRBD.
  The http queries were/are not the problem, the problem is that the
  controller then needs to get the state of all satellites.