invalidate cache on early boot
When the system boots (or the LINSTOR controller restarts) we might have
a situation where it report 0 for the sizes.

We want to differentiate a 0 from undef, because for undef we want to
report "inactive" (by returning undef). But if 0, we just want to skip
the cache and query current information. Otherwise people would need to
live with 0/inactive storage for the whole cache period.