show diskless disks in "VM Disks"
So far diskless assignments which are "temporary" as they are created
and deleted on the fly have not been shown in "VM Disks". This was
mainly because of how storage accounting worked previously. Even if
temporary the disk is there, so let's show it.

A long time ago storage accounting was based on the (LINSTOR) storage
pool. That was even before LINSTOR resource groups existed. This code
even got uglier when RGs started to consist of multiple storage pools.

For a long time the only handle we need is the name of the resource
group, we don't care about storage pools anymore. Remove all the storage
pool specific code and use what actually counts as a filter, namely the
RG name.