allow reassigning disks
This allows reassigning disks from one VM to another. For this to work
the plugin has to add support for renaming volumes, which is not that
easy as it sounds:

Renaming a LINSTOR/DRBD resource is basically out of the question
because of the limitations LINSTOR and DRBD have when it comes to
renaming resources. There are several options how renaming could be
done, with their own pros and cons:

Alternative 1: Using UUIDs as disk names. Usually disk names encode
their VMID in their name, as in vm-100-disk-1 (where 100 is the VMID,
the name "disk-1" does not really matter). If we use UUIDs on PVE and
LINSTOR/DRBD level, then we have to store the VMID somewhere else. This
could be a property on the LINSTOR RD level. This works fine in general,
but we have to do a "DB lookup" over the network whenever we need to
query the VMID. Especially "parse_volname" (which is part of the API)
makes this complicated as it is the only API function that does not get
a reference to it's storage configuration. Without that, we don't even
know the IP of the LINSTOR controller. This would require an API change
and then we still would have to implement another persistent cache
because that function is called in bursts. Even worse, the volume could
be renamed behind our back on a different node, so cache consistency is
thing.

Alternative 2: keep using the legacy vm-VMID-disk-NR names on the PVE
side, but use UUIDs on LINSTOR/DRBD level. We can then still use
"parse_volname" as is, as the VMID can still be parsed from the
"virtual" name. We can store the mapping on controller level (e.g.,
vm-100-disk-1 -> pm-$UUID). Disadvantage: we need to query the
controller whenever we need to know the actual LINSTOR/DRBD name, this
resulted in pretty ugly code and in pretty bad user experience. Whenever
a user/admin needs to fix stuff in LINSTOR/DRBD level they have to jump
through a layer virtual name resolution. Keeping names
current/consistent is another problem.

Solution: A mix of both. We use a (short) UUID as actual LINSTOR/DRBD
name, like pm-12345678, and annotate that with the VMID when used in PVE
as pm-12345678_$VMID. This has a lot of advantages:
- we can keep "parse_volname" as is, because the PVE name still contains
  the VMID, we can use a modified regex.
- mapping PVE names to actual LINSTOR names is trivial, just cut the
  "_$VMID" from the PVE name. It is trivial in code and it is trivial
  for human admins, the relation between PVE names and LINSTOR/DRBD
  names is obvious.
- We only need to update the VMID LINSTOR property when we do a rename,
  basically no "DB lookups".
- We can keep the mapping on RD level, when we delete the volume/RD, the
  mapping is automatically gone, no complexity in keeping the mappings
  consistent on controller level.

In theory there is a disadvantage of Alternative 1 and the Solution: The
rename API function can get a destination name. In that case we are not
able to actually rename the volume to some arbitrary name. We can only
"rename" the VMID part. This is usually - especially for reassigning
disks - fine as we don't get a fixed name but can generate one. In our
case we just keep the old/real name, but update the "virtual" VMID part.