fix cloud-init handling
This actually fixes two things:

By being pretty strict with names, we got too strict with cloud init
names. If a user wants to store their cloud images on DRBD, we currently
refuse these names (i.e., vm-ID-cloudinit). Add that to the name
checking part.

Second, PVE tries to be clever and recreates cloud-init images if it can
not stat the image on the node that starts the VM. In our case the stat
might fail because we usually create the diskless assignment in
"activate_volume()". But now PVE being clever tries to "alloc_image()"
and so far we failed because the image already exists on LINSTOR level.
If it already exists on LINSTOR level, just return and wait for the
actual "activate_volume()" to happen.