fix backups with EFI disks
A minimum size in general makes sense, if a disk is too small
drbd/drbd-utils simply refuse to work on these. The limit was 5MB, but
with that we saw weird side effects: on LVM for example 5MB became 8MB
(2 * 4MB). So far so good. When one backups a VM with an EFI disk on
DRBD everything looks fine, but on restore:

- PVE tries to create a 540672 bytes EFI disk
- the actual disk became larger (8MB in the example)
- vma has code that tolerates 4MB size differences, otherwise it bails
  out that sizes do not match.
- with 8MB we were over that tolerance => use a smaller cap (i.e. 3MB)
  to be in that tolerance

3MB is big enough so that none of our tools (LINSTOR/DRBD/drbd-utils)
complains, but small enough so that things work on LVM and ZFS.

This also fixes backups that have been made with "too large" EFI disks,
The important part is on restore and there the fix described above
applies.