keep tiebreaker
When LINSTOR deletes a diskless resource it converts it to a tiebreaker.
But if that is then not promoted, which would convert it again to a
diskless, but gets a second delete, the tiebreaker gets deleted because
LINSTOR has to assume that the deletion of the tiebreaker was
intentional. To avoid that and to always keep the tiebreaker around, set
this extra parameter. This does not break old controllers, but requires
LINSTOR >= 1.28.0 to have any effects.